moved HTTP handlers to a different file. queuing in progress
